src/main/java/com/ecommerce/app/ecommercebackend/service/ProductService.java
package com.ecommerce.app.ecommercebackend.service;

import com.ecommerce.app.ecommercebackend.api.repository.ProductRepository;
import com.ecommerce.app.ecommercebackend.exception.ApiResponseFailureException;
import com.ecommerce.app.ecommercebackend.model.Product;
import com.ecommerce.app.ecommercebackend.validation.FailureType;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import java.util.Optional;

import java.util.List;

@Service
public class ProductService {
    private ProductRepository productRepository;

    @Autowired
    public ProductService(ProductRepository productRepository) {
        this.productRepository = productRepository;
    }

    public List<Product> getProductList(){
        return productRepository.findAll();
    }

    public Product getProductById(Long id){
        Optional<Product> opProduct = productRepository.findById(id);
        return opProduct.orElseThrow(() -> new ApiResponseFailureException(FailureType.PRODUCT_NOT_FOUND, "Product could not be found for id " + id));
    }
}
